Get gene expression values from list of expression matrices.

Returns single vector of gene values across all datasets in list provided. Data can be in raw, normalized or scaled form. If matrices are in cell x gene format, set use.cols = TRUE.

getGeneValues(
  list,
  gene,
  use.cols = FALSE,
  methylation.indices = NULL,
  log2scale = FALSE,
  scale.factor = 10000
)

Arguments

list

List of gene x cell (or cell x gene) matrices

gene

Gene for which to return values (if gene is not found in appropriate dimnames will return vector of NA).

use.cols

Whether to query columns for desired gene (set to TRUE if matrices are cell x gene) (default FALSE).

methylation.indices

Indices of datasets with methylation data (never log2scaled) (default NULL).

log2scale

Whether to log2+1 scale (with multiplicative factor) values (default FALSE).

scale.factor

Scale factor to use with log2 scaling (default 10000).

Value

Plots to console (1-2 pages per factor)

Examples

if (FALSE) {
# liger object with factorization complete
# ligerex
gene_values <- getGeneValues(ligerex@raw.data, 'MALAT1')
}