Louvain algorithm for community detection

After quantile normalization, users can additionally run the Louvain algorithm for community detection, which is widely used in single-cell analysis and excels at merging small clusters into broad cell classes.

While using quantile normalized factor loadings (result from quantileNorm) is recommended, this function looks for unnormalized factor loadings (result from optimizeALS or online_iNMF) when the former is not available.

louvainCluster(
  object,
  resolution = 1,
  nNeighbors = 20,
  prune = 1/15,
  eps = 0.1,
  nRandomStarts = 10,
  nIterations = 100,
  useDims = NULL,
  groupSingletons = TRUE,
  clusterName = "louvain_cluster",
  seed = 1,
  verbose = TRUE,
  k = nNeighbors,
  dims.use = NULL,
  random.seed = 1
)

Arguments

object

liger object. Should run quantileNorm before calling.

resolution

Numeric, value of the resolution parameter, a larger value results in a larger number of communities with smaller sizes. Default 1.0.

nNeighbors

Integer, the maximum number of nearest neighbors to compute. Default 20.

prune

Numeric. Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the stringency of pruning. 0 for no pruning, while 1 prunes everything. Default 1/15.

eps

Numeric, the error bound of the nearest neighbor search. Default 0.1.

nRandomStarts

Integer number of random starts. Default 10.

nIterations

Integer, maximal number of iterations per random start. Default 100.

useDims

Indices of factors to use for Louvain clustering. Default NULL uses all available factors.

groupSingletons

Whether to group single cells that make up their own cluster in with the cluster they are most connected to. Default TRUE, if FALSE, assign all singletons to a "singleton" group.

clusterName

Name of the variable that will store the clustering result in cellMeta slot of object. Default "louvain_cluster".

seed

Seed of the random number generator. Default 1.

verbose

Logical. Whether to show information of the progress. Default TRUE.

k, dims.use, random.seed

Deprecated. See Usage section for replacement.

Value

object with refined cluster assignment updated in louvain_cluster variable in cellMeta slot. Can be fetched with object$louvain_cluster

Examples

data("pbmcPlot", package = "rliger")
pbmc <- louvainCluster(pbmcPlot)
#> Mon Mar 13 13:02:46 2023 ... Louvain clustering on quantile normalized cell factor loadings...