Perform iNMF on scaled datasets, and include unshared, scaled and normalized, features

Perform iNMF on scaled datasets, and include unshared, scaled and normalized, features

optimize_UANLS(
  object,
  k = 30,
  lambda = 5,
  max.iters = 30,
  nrep = 1,
  thresh = 1e-10,
  rand.seed = 1,
  print.obj = FALSE,
  vectorized.lambda = FALSE
)

Arguments

object

liger object. Should normalize, select genes, and scale before calling.

k

Inner dimension of factorization (number of factors).

lambda

The lambda penalty. Default 5

max.iters

Maximum number of block coordinate descent iterations to perform (default 30).

nrep

Number of restarts to perform (iNMF objective function is non-convex, so taking the best objective from multiple successive initializations is recommended). For easier reproducibility, this increments the random seed by 1 for each consecutive restart, so future factorizations of the same dataset can be run with one rep if necessary. (default 1)

thresh

Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh. (default 1e-6)

rand.seed

Random seed to allow reproducible results (default 1).

print.obj

Print objective function values after convergence (default FALSE).

vectorized.lamba

Whether or not to expect a vectorized lambda parameter ##########################################################################