[Deprecated] See downsample

This function mainly aims at downsampling datasets to a size suitable for plotting.

Usage

readSubset(
  object,
  slot.use = "normData",
  balance = NULL,
  max.cells = 1000,
  chunk = 1000,
  datasets.use = NULL,
  genes.use = NULL,
  rand.seed = 1,
  verbose = getOption("ligerVerbose", TRUE)
)

Arguments

object

liger object

slot.use

Only create subset from one or more of "rawData", "normData" and "scaleData". Default NULL subsets the whole object including downstream results.

balance

"all" for sampling maxCells cells from all datasets specified by useDatasets. "cluster" for sampling maxCells cells per cluster per dataset. "dataset" for maxCells cells per dataset.

max.cells

Max number of cells to sample from the grouping based on balance.

chunk

Integer. Number of maximum number of cells in each chunk, Default 1000.

datasets.use

Index selection of datasets to consider. Default NULL for using all datasets.

genes.use

Character vector. Subset features to this specified range. Default NULL does not subset features.

rand.seed

Random seed for reproducibility. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") or TRUE if users have not set.

Value

Subset of liger

object.

See also