Get cell metadata variable

Get cell metadata variable

# S3 method for liger
[[(x, i, ...)

Arguments

x

A liger object

i

Name or numeric index of cell meta data to fetch

...

Anything that S4Vectors::DataFrame method allows.

Value

If i is given, the selected metadata will be returned; if it is missing, the whole cell metadata table in S4Vectors::DataFrame class will be returned.

Examples

# Retrieve whole cellMeta
pbmc[[]]
#> DataFrame with 600 rows and 7 columns
#>                        dataset               barcode      nUMI     nGene
#>                       <factor>           <character> <numeric> <integer>
#> ctrl_AAACATACCTCGCT.1     ctrl ctrl_AAACATACCTCGCT.1      2151       102
#> ctrl_AAACGGCTCTTCGC.1     ctrl ctrl_AAACGGCTCTTCGC.1      1916       103
#> ctrl_AACACTCTAAGTAG.1     ctrl ctrl_AACACTCTAAGTAG.1      1869        95
#> ctrl_AACCGCCTCAGGAG.1     ctrl ctrl_AACCGCCTCAGGAG.1       733        84
#> ctrl_AACGTTCTTCCGTC.1     ctrl ctrl_AACGTTCTTCCGTC.1       573        76
#> ...                        ...                   ...       ...       ...
#> stim_TTCATGACTTATCC.1     stim stim_TTCATGACTTATCC.1      1091       101
#> stim_TTCATGACTTCAGG.1     stim stim_TTCATGACTTCAGG.1       431        86
#> stim_TTCGGAGATTTCAC.1     stim stim_TTCGGAGATTTCAC.1      1344       102
#> stim_TTGACACTTCCTGC.1     stim stim_TTGACACTTCCTGC.1      1423       105
#> stim_TTTGCATGAACGAA.1     stim stim_TTTGCATGAACGAA.1      2818       120
#>                            mito      ribo      hemo
#>                       <numeric> <numeric> <numeric>
#> ctrl_AAACATACCTCGCT.1         0  10.83217         0
#> ctrl_AAACGGCTCTTCGC.1         0  20.25052         0
#> ctrl_AACACTCTAAGTAG.1         0   5.08293         0
#> ctrl_AACCGCCTCAGGAG.1         0  42.70123         0
#> ctrl_AACGTTCTTCCGTC.1         0  38.04538         0
#> ...                         ...       ...       ...
#> stim_TTCATGACTTATCC.1         0   8.61595         0
#> stim_TTCATGACTTCAGG.1         0  41.29930         0
#> stim_TTCGGAGATTTCAC.1         0   8.85417         0
#> stim_TTGACACTTCCTGC.1         0  11.94659         0
#> stim_TTTGCATGAACGAA.1         0   9.43932         0
# Retrieve a variable
pbmc[["dataset"]]
#>   [1] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [16] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [31] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [46] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [61] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [76] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#>  [91] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [106] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [121] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [136] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [151] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [166] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [181] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [196] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [211] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [226] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [241] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [256] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [271] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [286] ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl ctrl
#> [301] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [316] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [331] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [346] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [361] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [376] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [391] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [406] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [421] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [436] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [451] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [466] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [481] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [496] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [511] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [526] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [541] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [556] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [571] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> [586] stim stim stim stim stim stim stim stim stim stim stim stim stim stim stim
#> Levels: ctrl stim